Refactored, introduced POFile and POEntry
