+++ pyproject.toml
@@ -38,7 +38,7 @@ classifiers = [
 
 [tool.poetry.scripts]
 manim_translate = 'manim_voiceover.translate:main'
-manim_render_translations = 'manim_voiceover.render_translations:main'
+manim_render_translations = 'manim_voiceover.translate.render:main'
 
 [tool.poetry.dependencies]
 python = ">=3.8,<4"
